Layoutコンポーネントでのレイアウト調整をuseLayoutフックに切り出し、コードを整理
components/Layout.tsx
- useState, useEffect, useRefをインポート
- headerHeight, barWidth, topHeight, bottomHeight, sidebarWidth, leftWidth, rightWidthをuseStateで定義
- useEffectで各値を計算してセットする
- sidebarHandlerDown, handleSidebarWidth, sidebarHandlerUp, leftRightHandlerDown, handleLeftWidth, leftRightHandlerUp, topBottomHandlerDown, handleTopBottomHeight, topBottomHandlerUpを定義
- useLayoutフックから各値とハンドラーを取得して利用するように修正

hooks/useLayout.tsx
- useState, useEffectをインポート
- headerHeight, barWidth, topHeight, bottomHeight, sidebarWidth, leftWidth, rightWidthをuseStateで定義
- useEffectで各値を計算してセットする
- sidebarHandlerDown, handleSidebarWidth, sidebarHandlerUp, leftRightHandlerDown, handleLeftWidth, leftRightHandlerUp, topBottomHandlerDown, handleTopBottomHeight, topBottomHandlerUpを定義
- headerHeight, barWidth, topHeight, bottomHeight, sidebarWidth, leftWidth, rightWidth, sidebarHandlerDown, leftRightHandlerDown, topBottomHandlerDownをエクスポート
